Split more tests
